relocate HashExtensionKat to Bodu.Security.Cryptography
CLAUDE.md grouped HashExtensionKat under Bodu.IO.Hashing, but its sole
consumer is HashAlgorithmExtensionsKatTests in
Bodu.Security.Cryptography.Test. Moving it to that project per the
single-consumer rule.

Moved to Bodu.Security.Cryptography/test/Infrastructure/
HashExtensionKat.cs (namespace Bodu.Security.Cryptography.Infrastructure,
matching the AeadKat / BlockCipherKat / CryptoHashKat moves from the
earlier crypto commit).

Updated HashAlgorithmExtensionsKatTests.cs to use the new namespace.

Verified: dotnet build passes 0 errors.
